multi tile actors and trees

+++ engine/config.h
@@ -135,7 +135,7 @@ namespace Config
 	inline uint32_t unitsBodyMassPerUnitFluidConsumed;
 	inline uint32_t baseMoveCost;
 	inline uint32_t goUpMoveCost;
-	inline uint32_t wanderDistanceModifier;
+	inline float wanderDistanceModifier;
 	inline Step stepsToDelayBeforeTryingAgainToCompleteAnObjective;
 	inline uint32_t hitScaleModifier;
 	inline uint32_t hitAreaToBodyPartVolumeRatioForFatalStrikeToVitalArea;
@@ -319,7 +319,7 @@ namespace Config
 		unitsBodyMassPerUnitFluidConsumed = data["unitsBodyMassPerUnitFluidConsumed"].get<uint32_t>();
 		baseMoveCost = data["baseMoveCost"].get<uint32_t>();
 		goUpMoveCost = data["goUpMoveCost"].get<uint32_t>();
-		wanderDistanceModifier = data["wanderDistanceModifier"].get<uint32_t>();
+		wanderDistanceModifier = data["wanderDistanceModifier"].get<float>();
 		stepsToDelayBeforeTryingAgainToCompleteAnObjective = data["secondsToDelayBeforeTryingAgainToCompleteAnObjective"].get<uint32_t>() * stepsPerSecond;
 		hitScaleModifier = data["hitScaleModifier"].get<uint32_t>();
 		hitAreaToBodyPartVolumeRatioForFatalStrikeToVitalArea = data["hitAreaToBodyPartVolumeRatioForFatalStrikeToVitalArea"].get<float>();
